+++ CHANGELOG.md
@@ -8,12 +8,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
-### Removed
+### Changed
 
 - Starting from `v1.21.0` of semantic conventions,
   `go.opentelemetry.io/otel/semconv/{version}/httpconv`
   and `go.opentelemetry.io/otel/semconv/{version}/netconv`
-  packages will no longer be published.
+  packages will no longer be published. (#4145)
 
 ## [1.16.0/0.39.0] 2023-05-18